feat: Integração do código dos Juniors
