--- Aluno.java
@@ -1,14 +1,0 @@
-public class Aluno {
-    private String apelido;
-
-    public Aluno(String apelido){
-        this.apelido = apelido;
-    }
-
-    public String getApelido(){
-        return apelido;
-    }
-    public void setApelido(String apelido){
-        this.apelido = apelido;
-    }
-}
