+++ ByteCraft/backend/src/main/java/com/bytecraft/model/Sala.java
@@ -1,0 +1,24 @@
+package com.bytecraft.model;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Set;
+
+
+//Classe do Pablo, já adaptada.
+@Entity
+@Getter
+@Setter
+public class Sala {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private byte codigoUnico;
+
+    @OneToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    private Set<Aluno> alunos;
+}
